+++ src/data/spcReducers.ts
@@ -13,6 +13,7 @@ export enum ControlLineReducerId {
   ucl = 'ucl',
   stdDev = 'stdDev',
   variance = 'variance',
+  gaussianPeak = 'gaussianPeak',
 }
 
 type SpcReducer = (field: Field, subgroupSize: number) => FieldCalcs;
@@ -47,7 +48,7 @@ export const controlLineReducers: ControlLineReducer[] = [
   {
     id: ControlLineReducerId.lsl,
     name: 'LSL',
-    description: 'Draws the LCL (Lower Control Limit), marking the lower boundary of the control chart.',
+    description: 'Draws the LSL (Lower Specification Limit), marking the lower boundary of the specification range.',
     computed: false,
     isStandard: false,
     color: '#C4162A',
@@ -55,7 +56,7 @@ export const controlLineReducers: ControlLineReducer[] = [
   {
     id: ControlLineReducerId.usl,
     name: 'USL',
-    description: 'Draws the USL (Upper Control Limit), marking the upper boundary of the control chart.',
+    description: 'Draws the USL (Upper Specification Limit), marking the upper boundary of the specification range.',
     computed: false,
     isStandard: false,
     color: '#C4162A',
@@ -103,9 +104,17 @@ export const controlLineReducers: ControlLineReducer[] = [
   {
     id: ControlLineReducerId.ucl,
     name: 'UCL',
-    description: 'Draws the USL (Upper Control Limit), marking the upper boundary of the control chart.',
+    description: 'Draws the UCL (Upper Control Limit), marking the upper boundary of the control chart.',
     computed: true,
     isStandard: false,
     color: '#F2495C',
+  },
+  {
+    id: ControlLineReducerId.gaussianPeak,
+    name: 'Gaussian Peak (µ)',
+    description: 'Draws a line at the peak of the fitted Gaussian curve (µ).',
+    computed: true,
+    isStandard: false,
+    color: 'yellow',
   },
 ];